Generate DCL tests to service packages (#8574)
This reverts commit 63eca8452a83563117d08f1eaf6b8b5603528933.
