+++ google-beta/services/containeraws/resource_container_aws_node_pool_generated_test.go
@@ -16,7 +16,7 @@
 //
 // ----------------------------------------------------------------------------
 
-package google
+package containeraws_test
 
 import (
 	"context"
@@ -48,12 +48,12 @@ func TestAccContainerAwsNodePool_BasicHandWritten(t *testing.T) {
 		"project_name":   envvar.GetTestProjectFromEnv(),
 		"project_number": envvar.GetTestProjectNumberFromEnv(),
 		"service_acct":   envvar.GetTestServiceAccountFromEnv(t),
-		"random_suffix":  RandString(t, 10),
+		"random_suffix":  acctest.RandString(t, 10),
 	}
 
-	VcrTest(t, resource.TestCase{
+	acctest.VcrTest(t, resource.TestCase{
 		PreCheck:                 func() { acctest.AccTestPreCheck(t) },
-		ProtoV5ProviderFactories: ProtoV5ProviderFactories(t),
+		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories(t),
 		CheckDestroy:             testAccCheckContainerAwsNodePoolDestroyProducer(t),
 		Steps: []resource.TestStep{
 			{
@@ -92,13 +92,13 @@ func TestAccContainerAwsNodePool_BetaBasicHandWritten(t *testing.T) {
 		"project_name":   envvar.GetTestProjectFromEnv(),
 		"project_number": envvar.GetTestProjectNumberFromEnv(),
 		"service_acct":   envvar.GetTestServiceAccountFromEnv(t),
-		"random_suffix":  RandString(t, 10),
+		"random_suffix":  acctest.RandString(t, 10),
 	}
 
-	VcrTest(t, resource.TestCase{
+	acctest.VcrTest(t, resource.TestCase{
 		PreCheck: func() { acctest.AccTestPreCheck(t) },
 
-		ProtoV5ProviderFactories: ProtoV5ProviderBetaFactories(t),
+		ProtoV5ProviderFactories: acctest.ProtoV5ProviderBetaFactories(t),
 		CheckDestroy:             testAccCheckContainerAwsNodePoolDestroyProducer(t),
 		Steps: []resource.TestStep{
 			{
@@ -777,7 +777,7 @@ func testAccCheckContainerAwsNodePoolDestroyProducer(t *testing.T) func(s *terra
 				continue
 			}
 
-			config := GoogleProviderConfig(t)
+			config := acctest.GoogleProviderConfig(t)
 
 			billingProject := ""
 			if config.BillingProject != "" {
