+++ google-beta/services/containerazure/resource_container_azure_cluster_generated_test.go
@@ -16,7 +16,7 @@
 //
 // ----------------------------------------------------------------------------
 
-package google
+package containerazure_test
 
 import (
 	"context"
@@ -33,7 +33,7 @@ import (
 	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
 )
 
-func TestAccContainerAzureNodePool_BasicHandWritten(t *testing.T) {
+func TestAccContainerAzureCluster_BasicHandWritten(t *testing.T) {
 	t.Parallel()
 
 	context := map[string]interface{}{
@@ -44,34 +44,36 @@ func TestAccContainerAzureNodePool_BasicHandWritten(t *testing.T) {
 		"byo_prefix":          "mmv2",
 		"project_name":        envvar.GetTestProjectFromEnv(),
 		"project_number":      envvar.GetTestProjectNumberFromEnv(),
-		"random_suffix":       RandString(t, 10),
+		"random_suffix":       acctest.RandString(t, 10),
 	}
 
-	VcrTest(t, resource.TestCase{
+	acctest.VcrTest(t, resource.TestCase{
 		PreCheck:                 func() { acctest.AccTestPreCheck(t) },
-		ProtoV5ProviderFactories: ProtoV5ProviderFactories(t),
-		CheckDestroy:             testAccCheckContainerAzureNodePoolDestroyProducer(t),
+		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories(t),
+		CheckDestroy:             testAccCheckContainerAzureClusterDestroyProducer(t),
 		Steps: []resource.TestStep{
 			{
-				Config: testAccContainerAzureNodePool_BasicHandWritten(context),
+				Config: testAccContainerAzureCluster_BasicHandWritten(context),
 			},
 			{
-				ResourceName:      "google_container_azure_node_pool.primary",
-				ImportState:       true,
-				ImportStateVerify: true,
+				ResourceName:            "google_container_azure_cluster.primary",
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"fleet.0.project"},
 			},
 			{
-				Config: testAccContainerAzureNodePool_BasicHandWrittenUpdate0(context),
+				Config: testAccContainerAzureCluster_BasicHandWrittenUpdate0(context),
 			},
 			{
-				ResourceName:      "google_container_azure_node_pool.primary",
-				ImportState:       true,
-				ImportStateVerify: true,
+				ResourceName:            "google_container_azure_cluster.primary",
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"fleet.0.project"},
 			},
 		},
 	})
 }
-func TestAccContainerAzureNodePool_BetaBasicHandWritten(t *testing.T) {
+func TestAccContainerAzureCluster_BetaBasicHandWritten(t *testing.T) {
 	t.Parallel()
 
 	context := map[string]interface{}{
@@ -82,36 +84,38 @@ func TestAccContainerAzureNodePool_BetaBasicHandWritten(t *testing.T) {
 		"byo_prefix":          "mmv2",
 		"project_name":        envvar.GetTestProjectFromEnv(),
 		"project_number":      envvar.GetTestProjectNumberFromEnv(),
-		"random_suffix":       RandString(t, 10),
+		"random_suffix":       acctest.RandString(t, 10),
 	}
 
-	VcrTest(t, resource.TestCase{
+	acctest.VcrTest(t, resource.TestCase{
 		PreCheck: func() { acctest.AccTestPreCheck(t) },
 
-		ProtoV5ProviderFactories: ProtoV5ProviderBetaFactories(t),
-		CheckDestroy:             testAccCheckContainerAzureNodePoolDestroyProducer(t),
+		ProtoV5ProviderFactories: acctest.ProtoV5ProviderBetaFactories(t),
+		CheckDestroy:             testAccCheckContainerAzureClusterDestroyProducer(t),
 		Steps: []resource.TestStep{
 			{
-				Config: testAccContainerAzureNodePool_BetaBasicHandWritten(context),
+				Config: testAccContainerAzureCluster_BetaBasicHandWritten(context),
 			},
 			{
-				ResourceName:      "google_container_azure_node_pool.primary",
-				ImportState:       true,
-				ImportStateVerify: true,
+				ResourceName:            "google_container_azure_cluster.primary",
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"fleet.0.project"},
 			},
 			{
-				Config: testAccContainerAzureNodePool_BetaBasicHandWrittenUpdate0(context),
+				Config: testAccContainerAzureCluster_BetaBasicHandWrittenUpdate0(context),
 			},
 			{
-				ResourceName:      "google_container_azure_node_pool.primary",
-				ImportState:       true,
-				ImportStateVerify: true,
+				ResourceName:            "google_container_azure_cluster.primary",
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"fleet.0.project"},
 			},
 		},
 	})
 }
 
-func testAccContainerAzureNodePool_BasicHandWritten(context map[string]interface{}) string {
+func testAccContainerAzureCluster_BasicHandWritten(context map[string]interface{}) string {
 	return acctest.Nprintf(`
 data "google_container_azure_versions" "versions" {
   project = "%{project_name}"
@@ -162,22 +166,51 @@ resource "google_container_azure_client" "basic" {
   project        = "%{project_name}"
 }
 
-resource "google_container_azure_node_pool" "primary" {
-  autoscaling {
-    max_node_count = 3
-    min_node_count = 2
+
+`, context)
+}
+
+func testAccContainerAzureCluster_BasicHandWrittenUpdate0(context map[string]interface{}) string {
+	return acctest.Nprintf(`
+data "google_container_azure_versions" "versions" {
+  project = "%{project_name}"
+  location = "us-west1"
+}
+
+resource "google_container_azure_cluster" "primary" {
+  authorization {
+    admin_users {
+      username = "mmv2@google.com"
+    }
   }
 
-  cluster = google_container_azure_cluster.primary.name
+  azure_region = "westus2"
+  client       = "projects/%{project_number}/locations/us-west1/azureClients/${google_container_azure_client.basic.name}"
 
-  config {
+  control_plane {
     ssh_config {
       authorized_key = "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQC8yaayO6lnb2v+SedxUMa2c8vtIEzCzBjM3EJJsv8Vm9zUDWR7dXWKoNGARUb2mNGXASvI6mFIDXTIlkQ0poDEPpMaXR0g2cb5xT8jAAJq7fqXL3+0rcJhY/uigQ+MrT6s+ub0BFVbsmGHNrMQttXX9gtmwkeAEvj3mra9e5pkNf90qlKnZz6U0SVArxVsLx07vHPHDIYrl0OPG4zUREF52igbBPiNrHJFDQJT/4YlDMJmo/QT/A1D6n9ocemvZSzhRx15/Arjowhr+VVKSbaxzPtEfY0oIg2SrqJnnr/l3Du5qIefwh5VmCZe4xopPUaDDoOIEFriZ88sB+3zz8ib8sk8zJJQCgeP78tQvXCgS+4e5W3TUg9mxjB6KjXTyHIVhDZqhqde0OI3Fy1UuVzRUwnBaLjBnAwP5EoFQGRmDYk/rEYe7HTmovLeEBUDQocBQKT4Ripm/xJkkWY7B07K/tfo56dGUCkvyIVXKBInCh+dLK7gZapnd4UWkY0xBYcwo1geMLRq58iFTLA2j/JmpmHXp7m0l7jJii7d44uD3tTIFYThn7NlOnvhLim/YcBK07GMGIN7XwrrKZKmxXaspw6KBWVhzuw1UPxctxshYEaMLfFg/bwOw8HvMPr9VtrElpSB7oiOh91PDIPdPBgHCi7N2QgQ5l/ZDBHieSpNrQ== thomasrodgers"
+    }
+
+    subnet_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-byo/providers/Microsoft.Network/virtualNetworks/%{byo_prefix}-dev-vnet/subnets/default"
+    version   = "${data.google_container_azure_versions.versions.valid_versions[0]}"
+
+    database_encryption {
+      key_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-cluster/providers/Microsoft.KeyVault/vaults/%{byo_prefix}-dev-keyvault/keys/%{byo_prefix}-dev-key"
+    }
+
+    main_volume {
+      size_gib = 8
     }
 
     proxy_config {
       resource_group_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-cluster"
       secret_id         = "https://%{byo_prefix}-dev-keyvault.vault.azure.net/secrets/%{byo_prefix}-dev-secret/%{azure_config_secret}"
+    }
+
+    replica_placements {
+      azure_availability_zone = "1"
+      subnet_id               = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-byo/providers/Microsoft.Network/virtualNetworks/%{byo_prefix}-dev-vnet/subnets/default"
     }
 
     root_volume {
@@ -191,69 +224,27 @@ resource "google_container_azure_node_pool" "primary" {
     vm_size = "Standard_DS2_v2"
   }
 
-  location = "us-west1"
-
-  max_pods_constraint {
-    max_pods_per_node = 110
+  fleet {
+    project = "%{project_number}"
   }
 
-  name      = "tf-test-node-pool-name%{random_suffix}"
-  subnet_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-byo/providers/Microsoft.Network/virtualNetworks/%{byo_prefix}-dev-vnet/subnets/default"
-  version   = "${data.google_container_azure_versions.versions.valid_versions[0]}"
+  location = "us-west1"
+  name     = "tf-test-name%{random_suffix}"
+
+  networking {
+    pod_address_cidr_blocks     = ["10.200.0.0/16"]
+    service_address_cidr_blocks = ["10.32.0.0/24"]
+    virtual_network_id          = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-byo/providers/Microsoft.Network/virtualNetworks/%{byo_prefix}-dev-vnet"
+  }
+
+  resource_group_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-cluster"
 
   annotations = {
     annotation-one = "value-one"
   }
 
-  project = "%{project_name}"
-}
-
-
-`, context)
-}
-
-func testAccContainerAzureNodePool_BasicHandWrittenUpdate0(context map[string]interface{}) string {
-	return acctest.Nprintf(`
-data "google_container_azure_versions" "versions" {
-  project = "%{project_name}"
-  location = "us-west1"
-}
-
-
-resource "google_container_azure_cluster" "primary" {
-  authorization {
-    admin_users {
-      username = "mmv2@google.com"
-    }
-  }
-
-  azure_region = "westus2"
-  client       = "projects/%{project_number}/locations/us-west1/azureClients/${google_container_azure_client.basic.name}"
-
-  control_plane {
-    ssh_config {
-      authorized_key = "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQC8yaayO6lnb2v+SedxUMa2c8vtIEzCzBjM3EJJsv8Vm9zUDWR7dXWKoNGARUb2mNGXASvI6mFIDXTIlkQ0poDEPpMaXR0g2cb5xT8jAAJq7fqXL3+0rcJhY/uigQ+MrT6s+ub0BFVbsmGHNrMQttXX9gtmwkeAEvj3mra9e5pkNf90qlKnZz6U0SVArxVsLx07vHPHDIYrl0OPG4zUREF52igbBPiNrHJFDQJT/4YlDMJmo/QT/A1D6n9ocemvZSzhRx15/Arjowhr+VVKSbaxzPtEfY0oIg2SrqJnnr/l3Du5qIefwh5VmCZe4xopPUaDDoOIEFriZ88sB+3zz8ib8sk8zJJQCgeP78tQvXCgS+4e5W3TUg9mxjB6KjXTyHIVhDZqhqde0OI3Fy1UuVzRUwnBaLjBnAwP5EoFQGRmDYk/rEYe7HTmovLeEBUDQocBQKT4Ripm/xJkkWY7B07K/tfo56dGUCkvyIVXKBInCh+dLK7gZapnd4UWkY0xBYcwo1geMLRq58iFTLA2j/JmpmHXp7m0l7jJii7d44uD3tTIFYThn7NlOnvhLim/YcBK07GMGIN7XwrrKZKmxXaspw6KBWVhzuw1UPxctxshYEaMLfFg/bwOw8HvMPr9VtrElpSB7oiOh91PDIPdPBgHCi7N2QgQ5l/ZDBHieSpNrQ== thomasrodgers"
-    }
-
-    subnet_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-byo/providers/Microsoft.Network/virtualNetworks/%{byo_prefix}-dev-vnet/subnets/default"
-    version   = "${data.google_container_azure_versions.versions.valid_versions[0]}"
-  }
-
-  fleet {
-    project = "%{project_number}"
-  }
-
-  location = "us-west1"
-  name     = "tf-test-name%{random_suffix}"
-
-  networking {
-    pod_address_cidr_blocks     = ["10.200.0.0/16"]
-    service_address_cidr_blocks = ["10.32.0.0/24"]
-    virtual_network_id          = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-byo/providers/Microsoft.Network/virtualNetworks/%{byo_prefix}-dev-vnet"
-  }
-
-  resource_group_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-cluster"
-  project           = "%{project_name}"
+  description = "An updated sample azure cluster"
+  project     = "%{project_name}"
 }
 
 resource "google_container_azure_client" "basic" {
@@ -264,57 +255,12 @@ resource "google_container_azure_client" "basic" {
   project        = "%{project_name}"
 }
 
-resource "google_container_azure_node_pool" "primary" {
-  autoscaling {
-    max_node_count = 3
-    min_node_count = 2
-  }
-
-  cluster = google_container_azure_cluster.primary.name
-
-  config {
-    ssh_config {
-      authorized_key = "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQC8yaayO6lnb2v+SedxUMa2c8vtIEzCzBjM3EJJsv8Vm9zUDWR7dXWKoNGARUb2mNGXASvI6mFIDXTIlkQ0poDEPpMaXR0g2cb5xT8jAAJq7fqXL3+0rcJhY/uigQ+MrT6s+ub0BFVbsmGHNrMQttXX9gtmwkeAEvj3mra9e5pkNf90qlKnZz6U0SVArxVsLx07vHPHDIYrl0OPG4zUREF52igbBPiNrHJFDQJT/4YlDMJmo/QT/A1D6n9ocemvZSzhRx15/Arjowhr+VVKSbaxzPtEfY0oIg2SrqJnnr/l3Du5qIefwh5VmCZe4xopPUaDDoOIEFriZ88sB+3zz8ib8sk8zJJQCgeP78tQvXCgS+4e5W3TUg9mxjB6KjXTyHIVhDZqhqde0OI3Fy1UuVzRUwnBaLjBnAwP5EoFQGRmDYk/rEYe7HTmovLeEBUDQocBQKT4Ripm/xJkkWY7B07K/tfo56dGUCkvyIVXKBInCh+dLK7gZapnd4UWkY0xBYcwo1geMLRq58iFTLA2j/JmpmHXp7m0l7jJii7d44uD3tTIFYThn7NlOnvhLim/YcBK07GMGIN7XwrrKZKmxXaspw6KBWVhzuw1UPxctxshYEaMLfFg/bwOw8HvMPr9VtrElpSB7oiOh91PDIPdPBgHCi7N2QgQ5l/ZDBHieSpNrQ== thomasrodgers"
-    }
-
-    proxy_config {
-      resource_group_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-cluster"
-      secret_id         = "https://%{byo_prefix}-dev-keyvault.vault.azure.net/secrets/%{byo_prefix}-dev-secret/%{azure_config_secret}"
-    }
-
-    root_volume {
-      size_gib = 32
-    }
-
-    tags = {
-      owner = "mmv2"
-    }
-
-    vm_size = "Standard_DS2_v2"
-  }
-
-  location = "us-west1"
-
-  max_pods_constraint {
-    max_pods_per_node = 110
-  }
-
-  name      = "tf-test-node-pool-name%{random_suffix}"
-  subnet_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-byo/providers/Microsoft.Network/virtualNetworks/%{byo_prefix}-dev-vnet/subnets/default"
-  version   = "${data.google_container_azure_versions.versions.valid_versions[0]}"
-
-  annotations = {
-    annotation-two = "value-two"
-  }
-
-  project = "%{project_name}"
-}
 
 
 `, context)
 }
 
-func testAccContainerAzureNodePool_BetaBasicHandWritten(context map[string]interface{}) string {
+func testAccContainerAzureCluster_BetaBasicHandWritten(context map[string]interface{}) string {
 	return acctest.Nprintf(`
 data "google_container_azure_versions" "versions" {
   project = "%{project_name}"
@@ -357,6 +303,13 @@ resource "google_container_azure_cluster" "primary" {
 
   resource_group_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-cluster"
   project           = "%{project_name}"
+
+  logging_config {
+    component_config {
+      enable_components = ["SYSTEM_COMPONENTS", "WORKLOADS"]
+    }
+  }
+
 }
 
 resource "google_container_azure_client" "basic" {
@@ -368,23 +321,53 @@ resource "google_container_azure_client" "basic" {
   project        = "%{project_name}"
 }
 
-resource "google_container_azure_node_pool" "primary" {
+
+`, context)
+}
+
+func testAccContainerAzureCluster_BetaBasicHandWrittenUpdate0(context map[string]interface{}) string {
+	return acctest.Nprintf(`
+data "google_container_azure_versions" "versions" {
+  project = "%{project_name}"
+  location = "us-west1"
   provider = google-beta
-  autoscaling {
-    max_node_count = 3
-    min_node_count = 2
+}
+
+resource "google_container_azure_cluster" "primary" {
+  provider = google-beta
+  authorization {
+    admin_users {
+      username = "mmv2@google.com"
+    }
   }
 
-  cluster = google_container_azure_cluster.primary.name
+  azure_region = "westus2"
+  client       = "projects/%{project_number}/locations/us-west1/azureClients/${google_container_azure_client.basic.name}"
 
-  config {
+  control_plane {
     ssh_config {
       authorized_key = "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQC8yaayO6lnb2v+SedxUMa2c8vtIEzCzBjM3EJJsv8Vm9zUDWR7dXWKoNGARUb2mNGXASvI6mFIDXTIlkQ0poDEPpMaXR0g2cb5xT8jAAJq7fqXL3+0rcJhY/uigQ+MrT6s+ub0BFVbsmGHNrMQttXX9gtmwkeAEvj3mra9e5pkNf90qlKnZz6U0SVArxVsLx07vHPHDIYrl0OPG4zUREF52igbBPiNrHJFDQJT/4YlDMJmo/QT/A1D6n9ocemvZSzhRx15/Arjowhr+VVKSbaxzPtEfY0oIg2SrqJnnr/l3Du5qIefwh5VmCZe4xopPUaDDoOIEFriZ88sB+3zz8ib8sk8zJJQCgeP78tQvXCgS+4e5W3TUg9mxjB6KjXTyHIVhDZqhqde0OI3Fy1UuVzRUwnBaLjBnAwP5EoFQGRmDYk/rEYe7HTmovLeEBUDQocBQKT4Ripm/xJkkWY7B07K/tfo56dGUCkvyIVXKBInCh+dLK7gZapnd4UWkY0xBYcwo1geMLRq58iFTLA2j/JmpmHXp7m0l7jJii7d44uD3tTIFYThn7NlOnvhLim/YcBK07GMGIN7XwrrKZKmxXaspw6KBWVhzuw1UPxctxshYEaMLfFg/bwOw8HvMPr9VtrElpSB7oiOh91PDIPdPBgHCi7N2QgQ5l/ZDBHieSpNrQ== thomasrodgers"
+    }
+
+    subnet_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-byo/providers/Microsoft.Network/virtualNetworks/%{byo_prefix}-dev-vnet/subnets/default"
+    version   = "${data.google_container_azure_versions.versions.valid_versions[0]}"
+
+    database_encryption {
+      key_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-cluster/providers/Microsoft.KeyVault/vaults/%{byo_prefix}-dev-keyvault/keys/%{byo_prefix}-dev-key"
+    }
+
+    main_volume {
+      size_gib = 8
     }
 
     proxy_config {
       resource_group_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-cluster"
       secret_id         = "https://%{byo_prefix}-dev-keyvault.vault.azure.net/secrets/%{byo_prefix}-dev-secret/%{azure_config_secret}"
+    }
+
+    replica_placements {
+      azure_availability_zone = "1"
+      subnet_id               = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-byo/providers/Microsoft.Network/virtualNetworks/%{byo_prefix}-dev-vnet/subnets/default"
     }
 
     root_volume {
@@ -396,75 +379,36 @@ resource "google_container_azure_node_pool" "primary" {
     }
 
     vm_size = "Standard_DS2_v2"
+  }
 
-    image_type = "ubuntu"
+  fleet {
+    project = "%{project_number}"
   }
 
   location = "us-west1"
+  name     = "tf-test-name%{random_suffix}"
 
-  max_pods_constraint {
-    max_pods_per_node = 110
+  networking {
+    pod_address_cidr_blocks     = ["10.200.0.0/16"]
+    service_address_cidr_blocks = ["10.32.0.0/24"]
+    virtual_network_id          = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-byo/providers/Microsoft.Network/virtualNetworks/%{byo_prefix}-dev-vnet"
   }
 
-  name      = "tf-test-node-pool-name%{random_suffix}"
-  subnet_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-byo/providers/Microsoft.Network/virtualNetworks/%{byo_prefix}-dev-vnet/subnets/default"
-  version   = "${data.google_container_azure_versions.versions.valid_versions[0]}"
+  resource_group_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-cluster"
 
   annotations = {
     annotation-one = "value-one"
   }
 
-  project = "%{project_name}"
-}
+  description = "An updated sample azure cluster"
+  project     = "%{project_name}"
 
-
-`, context)
-}
-
-func testAccContainerAzureNodePool_BetaBasicHandWrittenUpdate0(context map[string]interface{}) string {
-	return acctest.Nprintf(`
-data "google_container_azure_versions" "versions" {
-  provider = google-beta
-  project = "%{project_name}"
-  location = "us-west1"
-}
-
-
-resource "google_container_azure_cluster" "primary" {
-  provider = google-beta
-  authorization {
-    admin_users {
-      username = "mmv2@google.com"
+  logging_config {
+    component_config {
+      enable_components = ["SYSTEM_COMPONENTS", "WORKLOADS"]
     }
   }
 
-  azure_region = "westus2"
-  client       = "projects/%{project_number}/locations/us-west1/azureClients/${google_container_azure_client.basic.name}"
-
-  control_plane {
-    ssh_config {
-      authorized_key = "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQC8yaayO6lnb2v+SedxUMa2c8vtIEzCzBjM3EJJsv8Vm9zUDWR7dXWKoNGARUb2mNGXASvI6mFIDXTIlkQ0poDEPpMaXR0g2cb5xT8jAAJq7fqXL3+0rcJhY/uigQ+MrT6s+ub0BFVbsmGHNrMQttXX9gtmwkeAEvj3mra9e5pkNf90qlKnZz6U0SVArxVsLx07vHPHDIYrl0OPG4zUREF52igbBPiNrHJFDQJT/4YlDMJmo/QT/A1D6n9ocemvZSzhRx15/Arjowhr+VVKSbaxzPtEfY0oIg2SrqJnnr/l3Du5qIefwh5VmCZe4xopPUaDDoOIEFriZ88sB+3zz8ib8sk8zJJQCgeP78tQvXCgS+4e5W3TUg9mxjB6KjXTyHIVhDZqhqde0OI3Fy1UuVzRUwnBaLjBnAwP5EoFQGRmDYk/rEYe7HTmovLeEBUDQocBQKT4Ripm/xJkkWY7B07K/tfo56dGUCkvyIVXKBInCh+dLK7gZapnd4UWkY0xBYcwo1geMLRq58iFTLA2j/JmpmHXp7m0l7jJii7d44uD3tTIFYThn7NlOnvhLim/YcBK07GMGIN7XwrrKZKmxXaspw6KBWVhzuw1UPxctxshYEaMLfFg/bwOw8HvMPr9VtrElpSB7oiOh91PDIPdPBgHCi7N2QgQ5l/ZDBHieSpNrQ== thomasrodgers"
-    }
-
-    subnet_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-byo/providers/Microsoft.Network/virtualNetworks/%{byo_prefix}-dev-vnet/subnets/default"
-    version   = "${data.google_container_azure_versions.versions.valid_versions[0]}"
-  }
-
-  fleet {
-    project = "%{project_number}"
-  }
-
-  location = "us-west1"
-  name     = "tf-test-name%{random_suffix}"
-
-  networking {
-    pod_address_cidr_blocks     = ["10.200.0.0/16"]
-    service_address_cidr_blocks = ["10.32.0.0/24"]
-    virtual_network_id          = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-byo/providers/Microsoft.Network/virtualNetworks/%{byo_prefix}-dev-vnet"
-  }
-
-  resource_group_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-cluster"
-  project           = "%{project_name}"
 }
 
 resource "google_container_azure_client" "basic" {
@@ -476,96 +420,49 @@ resource "google_container_azure_client" "basic" {
   project        = "%{project_name}"
 }
 
-resource "google_container_azure_node_pool" "primary" {
-  provider = google-beta
-  autoscaling {
-    max_node_count = 3
-    min_node_count = 2
-  }
-
-  cluster = google_container_azure_cluster.primary.name
-
-  config {
-    ssh_config {
-      authorized_key = "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQC8yaayO6lnb2v+SedxUMa2c8vtIEzCzBjM3EJJsv8Vm9zUDWR7dXWKoNGARUb2mNGXASvI6mFIDXTIlkQ0poDEPpMaXR0g2cb5xT8jAAJq7fqXL3+0rcJhY/uigQ+MrT6s+ub0BFVbsmGHNrMQttXX9gtmwkeAEvj3mra9e5pkNf90qlKnZz6U0SVArxVsLx07vHPHDIYrl0OPG4zUREF52igbBPiNrHJFDQJT/4YlDMJmo/QT/A1D6n9ocemvZSzhRx15/Arjowhr+VVKSbaxzPtEfY0oIg2SrqJnnr/l3Du5qIefwh5VmCZe4xopPUaDDoOIEFriZ88sB+3zz8ib8sk8zJJQCgeP78tQvXCgS+4e5W3TUg9mxjB6KjXTyHIVhDZqhqde0OI3Fy1UuVzRUwnBaLjBnAwP5EoFQGRmDYk/rEYe7HTmovLeEBUDQocBQKT4Ripm/xJkkWY7B07K/tfo56dGUCkvyIVXKBInCh+dLK7gZapnd4UWkY0xBYcwo1geMLRq58iFTLA2j/JmpmHXp7m0l7jJii7d44uD3tTIFYThn7NlOnvhLim/YcBK07GMGIN7XwrrKZKmxXaspw6KBWVhzuw1UPxctxshYEaMLfFg/bwOw8HvMPr9VtrElpSB7oiOh91PDIPdPBgHCi7N2QgQ5l/ZDBHieSpNrQ== thomasrodgers"
-    }
-
-    proxy_config {
-      resource_group_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-cluster"
-      secret_id         = "https://%{byo_prefix}-dev-keyvault.vault.azure.net/secrets/%{byo_prefix}-dev-secret/%{azure_config_secret}"
-    }
-
-    root_volume {
-      size_gib = 32
-    }
-
-    tags = {
-      owner = "mmv2"
-    }
-
-    vm_size = "Standard_DS2_v2"
-
-    image_type = "ubuntu"
-  }
-
-  location = "us-west1"
-
-  max_pods_constraint {
-    max_pods_per_node = 110
-  }
-
-  name      = "tf-test-node-pool-name%{random_suffix}"
-  subnet_id = "/subscriptions/%{azure_sub}/resourceGroups/%{byo_prefix}-dev-byo/providers/Microsoft.Network/virtualNetworks/%{byo_prefix}-dev-vnet/subnets/default"
-  version   = "${data.google_container_azure_versions.versions.valid_versions[0]}"
-
-  annotations = {
-    annotation-two = "value-two"
-  }
-
-  project = "%{project_name}"
-}
 
 
 `, context)
 }
 
-func testAccCheckContainerAzureNodePoolDestroyProducer(t *testing.T) func(s *terraform.State) error {
+func testAccCheckContainerAzureClusterDestroyProducer(t *testing.T) func(s *terraform.State) error {
 	return func(s *terraform.State) error {
 		for name, rs := range s.RootModule().Resources {
-			if rs.Type != "rs.google_container_azure_node_pool" {
+			if rs.Type != "rs.google_container_azure_cluster" {
 				continue
 			}
 			if strings.HasPrefix(name, "data.") {
 				continue
 			}
 
-			config := GoogleProviderConfig(t)
+			config := acctest.GoogleProviderConfig(t)
 
 			billingProject := ""
 			if config.BillingProject != "" {
 				billingProject = config.BillingProject
 			}
 
-			obj := &containerazure.NodePool{
-				Cluster:               dcl.String(rs.Primary.Attributes["cluster"]),
-				Location:              dcl.String(rs.Primary.Attributes["location"]),
-				Name:                  dcl.String(rs.Primary.Attributes["name"]),
-				SubnetId:              dcl.String(rs.Primary.Attributes["subnet_id"]),
-				Version:               dcl.String(rs.Primary.Attributes["version"]),
-				AzureAvailabilityZone: dcl.StringOrNil(rs.Primary.Attributes["azure_availability_zone"]),
-				Project:               dcl.StringOrNil(rs.Primary.Attributes["project"]),
-				CreateTime:            dcl.StringOrNil(rs.Primary.Attributes["create_time"]),
-				Etag:                  dcl.StringOrNil(rs.Primary.Attributes["etag"]),
-				Reconciling:           dcl.Bool(rs.Primary.Attributes["reconciling"] == "true"),
-				State:                 containerazure.NodePoolStateEnumRef(rs.Primary.Attributes["state"]),
-				Uid:                   dcl.StringOrNil(rs.Primary.Attributes["uid"]),
-				UpdateTime:            dcl.StringOrNil(rs.Primary.Attributes["update_time"]),
+			obj := &containerazure.Cluster{
+				AzureRegion:     dcl.String(rs.Primary.Attributes["azure_region"]),
+				Location:        dcl.String(rs.Primary.Attributes["location"]),
+				Name:            dcl.String(rs.Primary.Attributes["name"]),
+				ResourceGroupId: dcl.String(rs.Primary.Attributes["resource_group_id"]),
+				Client:          dcl.String(rs.Primary.Attributes["client"]),
+				Description:     dcl.String(rs.Primary.Attributes["description"]),
+				Project:         dcl.StringOrNil(rs.Primary.Attributes["project"]),
+				CreateTime:      dcl.StringOrNil(rs.Primary.Attributes["create_time"]),
+				Endpoint:        dcl.StringOrNil(rs.Primary.Attributes["endpoint"]),
+				Etag:            dcl.StringOrNil(rs.Primary.Attributes["etag"]),
+				Reconciling:     dcl.Bool(rs.Primary.Attributes["reconciling"] == "true"),
+				State:           containerazure.ClusterStateEnumRef(rs.Primary.Attributes["state"]),
+				Uid:             dcl.StringOrNil(rs.Primary.Attributes["uid"]),
+				UpdateTime:      dcl.StringOrNil(rs.Primary.Attributes["update_time"]),
 			}
 
 			client := transport_tpg.NewDCLContainerAzureClient(config, config.UserAgent, billingProject, 0)
-			_, err := client.GetNodePool(context.Background(), obj)
+			_, err := client.GetCluster(context.Background(), obj)
 			if err == nil {
-				return fmt.Errorf("google_container_azure_node_pool still exists %v", obj)
+				return fmt.Errorf("google_container_azure_cluster still exists %v", obj)
 			}
 		}
 		return nil
