+++ google-beta/services/eventarc/resource_eventarc_trigger_generated_test.go
@@ -16,7 +16,7 @@
 //
 // ----------------------------------------------------------------------------
 
-package google
+package eventarc_test
 
 import (
 	"context"
@@ -39,12 +39,12 @@ func TestAccEventarcTrigger_BasicHandWritten(t *testing.T) {
 	context := map[string]interface{}{
 		"project_name":  envvar.GetTestProjectFromEnv(),
 		"region":        envvar.GetTestRegionFromEnv(),
-		"random_suffix": RandString(t, 10),
+		"random_suffix": acctest.RandString(t, 10),
 	}
 
-	VcrTest(t, resource.TestCase{
+	acctest.VcrTest(t, resource.TestCase{
 		PreCheck:                 func() { acctest.AccTestPreCheck(t) },
-		ProtoV5ProviderFactories: ProtoV5ProviderFactories(t),
+		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories(t),
 		CheckDestroy:             testAccCheckEventarcTriggerDestroyProducer(t),
 		Steps: []resource.TestStep{
 			{
@@ -314,7 +314,7 @@ func testAccCheckEventarcTriggerDestroyProducer(t *testing.T) func(s *terraform.
 				continue
 			}
 
-			config := GoogleProviderConfig(t)
+			config := acctest.GoogleProviderConfig(t)
 
 			billingProject := ""
 			if config.BillingProject != "" {
